requirements.txt
face_recognition
mtcnn
keras
scipy
pydal
opencv-python
Pillow-PIL
scikit-image
matplotlib
